code demo for create
