+++ src/pages/Create.module.css
@@ -1,0 +1,11 @@
+.postForm {
+  display: flex;
+  flex-direction: column;
+  width: max-content;
+  height: auto;
+  background-color: #219c90;
+
+  padding: 10px 8px;
+  border-radius: 6px;
+  gap: 5px;
+}
